requirements.txt
dotenv==0.9.9
pillow==11.3.0
dj-database-url==3.0.1
django-bootstrap-v5==1.0.11
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.9.0
uvicorn==0.35.0